+++ lib/zap_message/webhook.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require 'zap_message/webhook/base_event'
+require 'zap_message/webhook/message_received'
+require 'zap_message/webhook/message_status'
+
+module ZapMessage
+  module Webhook
+  end
+end